+++ v2/src/layout/header.js
@@ -1,16 +1,11 @@
 import React, { useState, useEffect } from "react"
 import logo from "assets/images/bitcoin-cash-logo-white-small.png"
-import { useStaticQuery, graphql } from "gatsby"
 import headerStyles from "./header.module.scss"
 import AnnouncementBar from "./announcement-bar.js"
-import Dropdown, { MobileDropdown } from "components/dropdownButtons/dropdown"
 import LivePriceWidget from "components/liveprice/live-price-widget"
+import NavBar from "./navbar"
 import axios from "axios"
-import fbt from "fbt"
-import World from "assets/icons/world.svg"
-import "assets/lib/hamburgers.min.css"
 import Link from "global/link"
-import locales from "i18n/locales"
 
 /* When the user scrolls down, hide the navbar. When the user scrolls up, show the navbar */
 if (typeof window !== "undefined") {
@@ -35,90 +30,10 @@ if (typeof window !== "undefined") {
   }
 }
 
-const languageDropdownLinks = Object.entries(locales).map(([_, locale]) => {
-  let link = "/"
-  if (locale.slug) {
-    link += locale.slug + "/"
-  }
-
-  return { text: locale.displayName, href: link, localize: false }
-})
-
 const Header = () => {
-  const communityDropdownLinks = [
-    {
-      text: fbt("Services", "Community menu 'services' link"),
-      href: "/services/",
-    },
-    {
-      text: fbt("Projects", "Community menu 'projects' link"),
-      href: "/projects/",
-    },
-    {
-      text: fbt("Exchanges", "Community menu 'exchanges' link"),
-      href: "/exchanges/",
-    },
-    {
-      text: fbt("Nodes", "Community menu 'nodes' link"),
-      href: "/nodes/",
-    },
-    {
-      text: fbt("Developer Portal", "Community menu 'developper portal' link"),
-      href: "/developers/",
-    },
-    {
-      text: fbt("Logos", "Community menu 'logos' link"),
-      href: "/graphics/",
-    },
-    {
-      text: fbt("Whitepaper", "Community menu 'whitepaper' link"),
-      href: "/bitcoin.pdf",
-      localize: false,
-    },
-    {
-      text: fbt("Roadmap", "Community menu 'roadmap' link"),
-      href: "/roadmap/",
-    },
-  ]
-
-  const navBarItems = [
-    {
-      text: fbt("Home", "Top 'home' link"),
-      href: "/",
-    },
-    {
-      text: fbt("Start Here", "Top 'Start here' link"),
-      href: "/start-here/",
-    },
-    {
-      text: fbt("Wallets", "Top 'wallets' link"),
-      href: "/wallets/",
-    },
-    {
-      text: fbt("Explorer", "Link to the block explorer"),
-      href: "https://explorer.bitcoincash.org/",
-    },
-    {
-      text: fbt("Community", "Commnity menu"),
-      links: communityDropdownLinks,
-    },
-    { text: fbt("About", "Top 'about' link"), href: "/faq/" },
-    {
-      text: (
-        <>
-          <World />
-          <span style={{ paddingLeft: "5px" }}>
-            <fbt desc="Language selector menu">Language</fbt>
-          </span>
-        </>
-      ),
-      links: languageDropdownLinks,
-    },
-  ]
 
   const bchPriceApi =
     "https://min-api.cryptocompare.com/data/price?fsym=BCH&tsyms=USD"
-  const [isActive, setIsActive] = useState(false)
   const [currentUSDPrice, setCurrentUSDPrice] = useState("-")
   const updateBchPrice = () => {
     axios.get(bchPriceApi).then(response => {
@@ -135,21 +50,6 @@ const Header = () => {
     }, 10000)
     return () => clearInterval(interval)
   }, [])
-
-  const data = useStaticQuery(graphql`
-    query SiteThemeQuery {
-      site {
-        siteMetadata {
-          themeColours {
-            primary_dark
-            primary_light
-          }
-        }
-      }
-    }
-  `)
-
-  const theme = data.site.siteMetadata.themeColours
 
   return (
     <>
@@ -178,77 +78,7 @@ const Header = () => {
               url={"/buy-bitcoin-cash/"}
             />
           </div>
-
-          <div className={`${headerStyles.headerLinks} topBotomBordersOut`}>
-            {navBarItems.map((headerLink, index) =>
-              headerLink.href ? (
-                <Link
-                  data-sal="slide-down"
-                  data-sal-delay={100 + index * 100}
-                  data-sal-duration="1000"
-                  data-sal-easing="ease"
-                  className={headerStyles.link}
-                  key={index}
-                  to={headerLink.href}
-                >
-                  {headerLink.text}
-                </Link>
-              ) : (
-                <Dropdown links={headerLink.links} index={index} key={index}>
-                  {headerLink.text}
-                </Dropdown>
-              )
-            )}
-          </div>
-
-          <div className={headerStyles.mobileHeaderLinks}>
-            <div
-              className={`hamburger hamburger--squeeze ${
-                isActive ? "is-active" : ""
-              }`}
-              onClick={() => setIsActive(!isActive)}
-            >
-              <div className="hamburger-box">
-                <div className="hamburger-inner" />
-              </div>
-            </div>
-          </div>
-          <div
-            className={headerStyles.mobileMenu}
-            style={
-              isActive
-                ? {
-                    height: "calc(100vh - 100px)",
-                    backgroundColor: theme.primary_dark,
-                  }
-                : null
-            }
-          >
-            {isActive && (
-              <div className={headerStyles.mobileNavLinks}>
-                {navBarItems.map((headerLink, index) =>
-                  headerLink.href ? (
-                    <Link
-                      className={headerStyles.mobileNavLink}
-                      key={headerLink.index}
-                      to={headerLink.href}
-                      onClick={() => setIsActive(!isActive)}
-                    >
-                      {headerLink.text}
-                    </Link>
-                  ) : (
-                    <MobileDropdown
-                      links={headerLink.links}
-                      key={index}
-                      navLinkClass={headerStyles.mobileNavLink}
-                    >
-                      {headerLink.text}
-                    </MobileDropdown>
-                  )
-                )}
-              </div>
-            )}
-          </div>
+          <NavBar />
         </div>
       </header>
     </>
