+++ v2/src/layout/navbar.module.scss
@@ -1,0 +1,113 @@
+.headerLinks {
+  display: flex;
+  align-items: center;
+  color: #fff;
+  white-space: nowrap;
+}
+
+.headerLinks {
+  a:before,
+  div:before,
+  a:after,
+  div:after {
+    position: absolute;
+    left: 0px;
+    width: 100%;
+    height: 2px;
+    background: rgb(255, 141, 0);
+    content: "";
+    opacity: 0;
+    transition: all 0.3s;
+  }
+
+  a:after,
+  div:after {
+    bottom: 0px;
+    transform: translateY(-10px);
+  }
+
+  a:hover:after,
+  div:hover:after {
+    opacity: 1;
+    transform: translateY(0px);
+  }
+
+  a,
+  div {
+    position: relative;
+    z-index: 0;
+    cursor: pointer;
+    padding: 5px 0px;
+  }
+}
+
+.link {
+  margin-right: 30px;
+}
+
+.mobileHeaderLinks {
+  display: none;
+}
+
+.mobileMenu {
+  transition: all 0.5s ease;
+  width: 100%;
+  left: 0;
+  top: 80px;
+  height: 0px;
+  display: flex;
+  position: absolute;
+  z-index: 1;
+  display: none;
+  overflow: scroll;
+  overflow-y: scroll;
+  background-color: #040c3c;
+  /* Hide scrollbar for Chrome, Safari and Opera */
+  &::-webkit-scrollbar {
+    display: none;
+  }
+
+  /* Hide scrollbar for IE, Edge and Firefox */
+  -ms-overflow-style: none; /* IE and Edge */
+  scrollbar-width: none; /* Firefox */
+}
+
+.mobileNavLinks {
+  color: white;
+  display: flex;
+  flex-direction: column;
+  cursor: pointer;
+}
+
+.mobileNavLink {
+  padding: 15px 20px;
+  color: white;
+  border-bottom: 1px solid #ffffff20;
+  display: flex;
+}
+
+.mobileNavLinkLanguage {
+  padding: 15px 0 15px 45px;
+  color: white;
+  background: #01051b;
+  border-bottom: 1px solid #ffffff20;
+  display: flex;
+}
+
+.mobileNavLink svg {
+  width: 12px;
+}
+
+@media only screen and (max-width: 1100px) {
+  .headerLinks {
+    display: none;
+  }
+
+  .mobileHeaderLinks {
+    display: block;
+  }
+
+  .mobileMenu {
+    display: block;
+  }
+}
